+++ sdk/include/opentelemetry/sdk/metrics/exemplar/no_exemplar_reservoir.h
@@ -7,7 +7,7 @@
 #  include "opentelemetry/context/context.h"
 #  include "opentelemetry/nostd/shared_ptr.h"
 #  include "opentelemetry/sdk/common/attribute_utils.h"
-#  include "opentelemetry/sdk/metrics/examplar/reservoir.h"
+#  include "opentelemetry/sdk/metrics/exemplar/reservoir.h"
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace sdk
@@ -25,14 +25,16 @@ public:
 
   void OfferMeasurement(long value,
                         const MetricAttributes &attributes,
-                        const opentelemetry::context::Context &context) noexcept override
+                        const opentelemetry::context::Context &context,
+                        const opentelemetry::common::SystemTimestamp &timestamp) noexcept override
   {
     // Stores nothing
   }
 
   void OfferMeasurement(double value,
                         const MetricAttributes &attributes,
-                        const opentelemetry::context::Context &context) noexcept override
+                        const opentelemetry::context::Context &context,
+                        const opentelemetry::common::SystemTimestamp &timestamp) noexcept override
   {
     // Stores nothing.
   }
