+++ sdk/include/opentelemetry/sdk/metrics/state/multi_metric_storage.h
@@ -4,7 +4,6 @@
 #pragma once
 #ifndef ENABLE_METRICS_PREVIEW
 #  include "opentelemetry/common/key_value_iterable_view.h"
-#  include "opentelemetry/sdk/metrics/examplar/reservoir.h"
 #  include "opentelemetry/sdk/metrics/instruments.h"
 #  include "opentelemetry/sdk/metrics/state/metric_storage.h"
 
@@ -19,16 +18,11 @@ namespace metrics
 class MultiMetricStorage : public WritableMetricStorage
 {
 public:
-  MultiMetricStorage(nostd::shared_ptr<ExemplarReservoir> exemplar_reservoir)
-      : exemplar_reservoir_(exemplar_reservoir)
-  {}
-
   void AddStorage(std::shared_ptr<WritableMetricStorage> storage) { storages_.push_back(storage); }
 
   virtual void RecordLong(long value,
                           const opentelemetry::context::Context &context) noexcept override
   {
-    exemplar_reservoir_->OfferMeasurement(value, {}, context);
     for (auto &s : storages_)
     {
       s->RecordLong(value, context);
@@ -39,7 +33,6 @@ public:
                           const opentelemetry::common::KeyValueIterable &attributes,
                           const opentelemetry::context::Context &context) noexcept override
   {
-    exemplar_reservoir_->OfferMeasurement(value, attributes, context);
     for (auto &s : storages_)
     {
       s->RecordLong(value, attributes, context);
@@ -49,7 +42,6 @@ public:
   virtual void RecordDouble(double value,
                             const opentelemetry::context::Context &context) noexcept override
   {
-    exemplar_reservoir_->OfferMeasurement(value, {}, context);
     for (auto &s : storages_)
     {
       s->RecordDouble(value, context);
@@ -60,7 +52,6 @@ public:
                             const opentelemetry::common::KeyValueIterable &attributes,
                             const opentelemetry::context::Context &context) noexcept override
   {
-    exemplar_reservoir_->OfferMeasurement(value, attributes, context);
     for (auto &s : storages_)
     {
       s->RecordDouble(value, attributes, context);
@@ -69,7 +60,6 @@ public:
 
 private:
   std::vector<std::shared_ptr<WritableMetricStorage>> storages_;
-  nostd::shared_ptr<ExemplarReservoir> exemplar_reservoir_;
 };
 
 }  // namespace metrics
