+++ sdk/include/opentelemetry/sdk/metrics/state/sync_metric_storage.h
@@ -7,7 +7,7 @@
 #  include "opentelemetry/sdk/common/attributemap_hash.h"
 #  include "opentelemetry/sdk/instrumentationlibrary/instrumentation_library.h"
 #  include "opentelemetry/sdk/metrics/aggregation/default_aggregation.h"
-#  include "opentelemetry/sdk/metrics/examplar/reservoir.h"
+#  include "opentelemetry/sdk/metrics/exemplar/reservoir.h"
 #  include "opentelemetry/sdk/metrics/state/attributes_hashmap.h"
 #  include "opentelemetry/sdk/metrics/state/metric_storage.h"
 #  include "opentelemetry/sdk/metrics/view/attributes_processor.h"
@@ -48,7 +48,7 @@ public:
     {
       return;
     }
-    exemplar_reservoir_->OfferMeasurement(value, {}, context);
+    exemplar_reservoir_->OfferMeasurement(value, {}, context, std::chrono::system_clock::now());
     attributes_hashmap_->GetOrSetDefault({}, create_default_aggregation_)->Aggregate(value);
   }
 
@@ -61,7 +61,8 @@ public:
       return;
     }
 
-    exemplar_reservoir_->OfferMeasurement(value, attributes, context);
+    exemplar_reservoir_->OfferMeasurement(value, attributes, context,
+                                          std::chrono::system_clock::now());
     auto attr = attributes_processor_->process(attributes);
     attributes_hashmap_->GetOrSetDefault(attr, create_default_aggregation_)->Aggregate(value);
   }
@@ -73,7 +74,7 @@ public:
       return;
     }
 
-    exemplar_reservoir_->OfferMeasurement(value, {}, context);
+    exemplar_reservoir_->OfferMeasurement(value, {}, context, std::chrono::system_clock::now());
     attributes_hashmap_->GetOrSetDefault({}, create_default_aggregation_)->Aggregate(value);
   }
 
@@ -81,13 +82,15 @@ public:
                     const opentelemetry::common::KeyValueIterable &attributes,
                     const opentelemetry::context::Context &context) noexcept override
   {
-    exemplar_reservoir_->OfferMeasurement(value, attributes, context);
+    exemplar_reservoir_->OfferMeasurement(value, attributes, context,
+                                          std::chrono::system_clock::now());
     if (instrument_descriptor_.value_type_ != InstrumentValueType::kDouble)
     {
       return;
     }
 
-    exemplar_reservoir_->OfferMeasurement(value, attributes, context);
+    exemplar_reservoir_->OfferMeasurement(value, attributes, context,
+                                          std::chrono::system_clock::now());
     auto attr = attributes_processor_->process(attributes);
     attributes_hashmap_->GetOrSetDefault(attr, create_default_aggregation_)->Aggregate(value);
   }
