+++ sdk/test/metrics/examplar/CMakeLists.txt
@@ -1,4 +1,4 @@
-foreach(testname no_examplar_reservoir_test never_sample_filter_test
+foreach(testname no_exemplar_reservoir_test never_sample_filter_test
                  always_sample_filter_test)
   add_executable(${testname} "${testname}.cc")
   target_link_libraries(${testname} ${GTEST_BOTH_LIBRARIES}
