+++ sdk/test/metrics/examplar/always_sample_filter_test.cc
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #ifndef ENABLE_METRICS_PREVIEW
-#  include "opentelemetry/sdk/metrics/examplar/always_sample_filter.h"
+#  include "opentelemetry/sdk/metrics/exemplar/always_sample_filter.h"
 #  include <gtest/gtest.h>
 
 using namespace opentelemetry::sdk::metrics;
