+++ sdk/test/metrics/sync_instruments_test.cc
@@ -5,7 +5,7 @@
 #  include "opentelemetry/sdk/metrics/sync_instruments.h"
 #  include "opentelemetry/context/context.h"
 #  include "opentelemetry/sdk/instrumentationlibrary/instrumentation_library.h"
-#  include "opentelemetry/sdk/metrics/examplar/no_examplar_reservoir.h"
+#  include "opentelemetry/sdk/metrics/exemplar/no_exemplar_reservoir.h"
 #  include "opentelemetry/sdk/metrics/state/multi_metric_storage.h"
 
 #  include <gtest/gtest.h>
@@ -22,8 +22,7 @@ TEST(SyncInstruments, LongCounter)
 {
   InstrumentDescriptor instrument_descriptor = {
       "long_counter", "description", "1", InstrumentType::kCounter, InstrumentValueType::kLong};
-  std::unique_ptr<WritableMetricStorage> metric_storage(
-      new MultiMetricStorage(NoExemplarReservoir::GetNoExemplarReservoir()));
+  std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   LongCounter counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Add(10l));
   EXPECT_NO_THROW(counter.Add(10l, opentelemetry::context::Context{}));
@@ -42,8 +41,7 @@ TEST(SyncInstruments, DoubleCounter)
 {
   InstrumentDescriptor instrument_descriptor = {
       "double_counter", "description", "1", InstrumentType::kCounter, InstrumentValueType::kDouble};
-  std::unique_ptr<WritableMetricStorage> metric_storage(
-      new MultiMetricStorage(NoExemplarReservoir::GetNoExemplarReservoir()));
+  std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   DoubleCounter counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Add(10.10));
   EXPECT_NO_THROW(counter.Add(10.10, opentelemetry::context::Context{}));
@@ -63,8 +61,7 @@ TEST(SyncInstruments, LongUpDownCounter)
   InstrumentDescriptor instrument_descriptor = {"long_updowncounter", "description", "1",
                                                 InstrumentType::kUpDownCounter,
                                                 InstrumentValueType::kLong};
-  std::unique_ptr<WritableMetricStorage> metric_storage(
-      new MultiMetricStorage(NoExemplarReservoir::GetNoExemplarReservoir()));
+  std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   LongUpDownCounter counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Add(10l));
   EXPECT_NO_THROW(counter.Add(10l, opentelemetry::context::Context{}));
@@ -84,8 +81,7 @@ TEST(SyncInstruments, DoubleUpDownCounter)
   InstrumentDescriptor instrument_descriptor = {"double_updowncounter", "description", "1",
                                                 InstrumentType::kUpDownCounter,
                                                 InstrumentValueType::kDouble};
-  std::unique_ptr<WritableMetricStorage> metric_storage(
-      new MultiMetricStorage(NoExemplarReservoir::GetNoExemplarReservoir()));
+  std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   DoubleUpDownCounter counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Add(10.10));
   EXPECT_NO_THROW(counter.Add(10.10, opentelemetry::context::Context{}));
@@ -104,8 +100,7 @@ TEST(SyncInstruments, LongHistogram)
 {
   InstrumentDescriptor instrument_descriptor = {
       "long_histogram", "description", "1", InstrumentType::kHistogram, InstrumentValueType::kLong};
-  std::unique_ptr<WritableMetricStorage> metric_storage(
-      new MultiMetricStorage(NoExemplarReservoir::GetNoExemplarReservoir()));
+  std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   LongHistogram counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Record(10l, opentelemetry::context::Context{}));
   EXPECT_NO_THROW(counter.Record(10l, opentelemetry::context::Context{}));
@@ -122,8 +117,7 @@ TEST(SyncInstruments, DoubleHistogram)
   InstrumentDescriptor instrument_descriptor = {"double_histogram", "description", "1",
                                                 InstrumentType::kHistogram,
                                                 InstrumentValueType::kDouble};
-  std::unique_ptr<WritableMetricStorage> metric_storage(
-      new MultiMetricStorage(NoExemplarReservoir::GetNoExemplarReservoir()));
+  std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   DoubleHistogram counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Record(10.10, opentelemetry::context::Context{}));
   EXPECT_NO_THROW(counter.Record(10.10, opentelemetry::context::Context{}));
